Hotfix/fix deployment

+++ Makefile
@@ -2,7 +2,9 @@ DC := docker compose
 D := docker
 
 .DEFAULT_GOAL := dev
+
 include .env
+include config.env
 .EXPORT_ALL_VARIABLES:
 
 .PHONY: check-env
@@ -11,6 +13,9 @@ check-env:
 
 .PHONY: dev
 dev: check-env
+	[ -n "$(DEV_HTTP_PORT)" ] && \
+	[ -n "$(DEV_HTTPS_PORT)" ] && \
+	[ -n "$(DEV_DOMAINS)" ] && \
 	HTTP_PORT=$(DEV_HTTP_PORT) \
 	HTTPS_PORT=$(DEV_HTTPS_PORT) \
 	DOMAINS=$(DEV_DOMAINS) \
@@ -18,6 +23,9 @@ dev: check-env
 
 .PHONY: prod
 prod: check-env
+	[ -n "$(PROD_HTTP_PORT)" ] && \
+	[ -n "$(PROD_HTTPS_PORT)" ] && \
+	[ -n "$(PROD_DOMAINS)" ] && \
 	HTTP_PORT=$(PROD_HTTP_PORT) \
 	HTTPS_PORT=$(PROD_HTTPS_PORT) \
 	DOMAINS=$(PROD_DOMAINS) \
@@ -29,7 +37,7 @@ down:
 
 .PHONY: clean
 clean: down
-	$(RM) backend/$(DB_PATH)
+	[ -n "$(DB_PATH)" ] && $(RM) backend/$(DB_PATH)
 	$(D) system prune -f
 
 .PHONY: deepclean
